Update for restoring Level-8 changes
