--- src/main/java/BobbyException.java
@@ -1,8 +1,0 @@
-/**
- * For all errors related to Bobby
- */
-public class BobbyException extends Exception{
-    public BobbyException(String message) {
-        super(message);
-    }
-}
