+++ src/main/java/bobby/task/Task.java
@@ -1,5 +1,11 @@
 package bobby.task;
 
+import bobby.exception.BobbyException;
+
+import java.time.LocalDateTime;
+import java.time.format.DateTimeFormatter;
+import java.time.format.DateTimeParseException;
+
 /**
  * abstract class that todo, deadline and event inherit from
  */
@@ -40,6 +46,39 @@ public abstract class Task {
      */
     public void unmark() {
         isMark = false;
+    }
+
+    /**
+     * parses the given string into datetime
+     *
+     * @param datetime string
+     * @return LocalDateTime object
+     */
+    public LocalDateTime parseString(String datetime) throws BobbyException {
+        try {
+            DateTimeFormatter formatter = DateTimeFormatter.ofPattern("yyyy-MM-dd HHmm");
+            return LocalDateTime.parse(datetime, formatter);
+        } catch (DateTimeParseException e) {
+            throw new BobbyException("by/from/to must be in the format yyyy-MM-dd HHmm");
+        } catch (NullPointerException e) {
+            throw new BobbyException("by/from/to cannot be left empty");
+        }
+    }
+
+    /**
+     * converts LocalDateTime into toString() friendly format
+     *
+     * @param datetime LocalDateTime object
+     * @return String
+     */
+    public String datetimeToString(LocalDateTime datetime) {
+        DateTimeFormatter formatter = DateTimeFormatter.ofPattern("dd/MM/yyyy HHmm");
+        return datetime.format(formatter);
+    }
+
+    public String datetimeToStorage(LocalDateTime datetime) {
+        DateTimeFormatter formatter = DateTimeFormatter.ofPattern("yyyy-MM-dd HHmm");
+        return datetime.format(formatter);
     }
 
     /**
