+++ src/main/java/bobby/task/TaskList.java
@@ -48,16 +48,16 @@ public class TaskList {
                 if (split.length == 2 && split[1].substring(0, 2).equalsIgnoreCase("by")) {
                     taskList.add(new Deadline(split[0], isMark, split[1].substring(3)));
                 } else {
-                    throw new BobbyException("This is not the right format for adding a task.");
+                    throw new BobbyException("Use the yyyy-MM-dd HHmm format.");
                 }
             } else {
                 if (split.length == 3 &&
-                    split[1].substring(0, 4).equalsIgnoreCase("from") &&
-                    split[2].substring(0, 2).equalsIgnoreCase("to")
+                        split[1].substring(0, 4).equalsIgnoreCase("from") &&
+                        split[2].substring(0, 2).equalsIgnoreCase("to")
                 ) {
                     taskList.add(new Event(split[0], isMark, split[1].substring(5), split[2].substring(3)));
                 } else {
-                    throw new BobbyException("This is not the right format for adding a task.");
+                    throw new BobbyException("Use the yyyy-MM-dd HHmm format.");
                 }
             }
         }
@@ -84,7 +84,7 @@ public class TaskList {
      */
     public void markTask(int taskNum) throws BobbyException {
         if (taskNum <= 0 || taskNum > taskList.size()) {
-            throw new BobbyException("You can't delete a task that does not exist");
+            throw new BobbyException("You can't mark a task that does not exist");
         }
         taskList.get(taskNum - 1).mark();
     }
@@ -97,7 +97,7 @@ public class TaskList {
      */
     public void unmarkTask(int taskNum) throws BobbyException {
         if (taskNum <= 0 || taskNum > taskList.size()) {
-            throw new BobbyException("You can't delete a task that does not exist");
+            throw new BobbyException("You can't unmark a task that does not exist");
         }
         taskList.get(taskNum - 1).unmark();
     }
