+++ text-ui-test/EXPECTED.TXT
@@ -13,30 +13,30 @@
     todo requires arguments.
     ________________________________________________________________
     ________________________________________________________________
-    Use this format: deadline {name} /by dd/MM/yyyy HHmm
+    Use the yyyy-MM-dd HHmm format.
     ________________________________________________________________
     ________________________________________________________________
-    by/from/to must be in the format dd/MM/yyyy HHmm
+    by/from/to must be in the format yyyy-MM-dd HHmm
     ________________________________________________________________
     ________________________________________________________________
     I've added this Deadline.
-       [D][ ] deadline1 (by: 01-01-2025 10:10)
+       [D][ ] deadline1 (by: 01/01/2025 1010)
     ________________________________________________________________
     ________________________________________________________________
-    Use this format: event {name} /from dd/MM/yyyy HHmm /to dd/MM/yyyy HHmm
+    Use the yyyy-MM-dd HHmm format.
     ________________________________________________________________
     ________________________________________________________________
-    by/from/to must be in the format dd/MM/yyyy HHmm
+    by/from/to must be in the format yyyy-MM-dd HHmm
     ________________________________________________________________
     ________________________________________________________________
     I've added this Event.
-       [E][ ] event1 (from: 01-01-2025 10:10 to: 02-01-2025 10:11)
+       [E][ ] event1 (from: 05/05/2025 1010 to: 10/06/2025 1011)
     ________________________________________________________________
     ________________________________________________________________
     Here are your tasks:
     1. [T][ ] task1
-    2. [D][ ] deadline1 (by: 01-01-2025 10:10)
-    3. [E][ ] event1 (from: 01-01-2025 10:10 to: 02-01-2025 10:11)
+    2. [D][ ] deadline1 (by: 01/01/2025 1010)
+    3. [E][ ] event1 (from: 05/05/2025 1010 to: 10/06/2025 1011)
     ________________________________________________________________
     ________________________________________________________________
     list should not have any arguments.
@@ -57,8 +57,8 @@
     ________________________________________________________________
     Here are your tasks:
     1. [T][X] task1
-    2. [D][ ] deadline1 (by: 01-01-2025 10:10)
-    3. [E][ ] event1 (from: 01-01-2025 10:10 to: 02-01-2025 10:11)
+    2. [D][ ] deadline1 (by: 01/01/2025 1010)
+    3. [E][ ] event1 (from: 05/05/2025 1010 to: 10/06/2025 1011)
     ________________________________________________________________
     ________________________________________________________________
     unmark requires arguments.
@@ -76,8 +76,8 @@
     ________________________________________________________________
     Here are your tasks:
     1. [T][ ] task1
-    2. [D][ ] deadline1 (by: 01-01-2025 10:10)
-    3. [E][ ] event1 (from: 01-01-2025 10:10 to: 02-01-2025 10:11)
+    2. [D][ ] deadline1 (by: 01/01/2025 1010)
+    3. [E][ ] event1 (from: 05/05/2025 1010 to: 10/06/2025 1011)
     ________________________________________________________________
     ________________________________________________________________
     delete requires arguments.
@@ -94,8 +94,8 @@
     ________________________________________________________________
     ________________________________________________________________
     Here are your tasks:
-    1. [D][ ] deadline1 (by: 01-01-2025 10:10)
-    2. [E][ ] event1 (from: 01-01-2025 10:10 to: 02-01-2025 10:11)
+    1. [D][ ] deadline1 (by: 01/01/2025 1010)
+    2. [E][ ] event1 (from: 05/05/2025 1010 to: 10/06/2025 1011)
     ________________________________________________________________
     ________________________________________________________________
     Bye. Hope to see you again soon!
